+++ src/main/java/pyre/coloredredstone/blocks/BlockColoredRedstone.java
@@ -135,4 +135,22 @@ public class BlockColoredRedstone extends Block implements IColoredFeatures, IBl
             }
         }
     }
+
+    @Override
+    public int getFlammability(IBlockAccess world, BlockPos pos, EnumFacing face) {
+        EnumColor color = getColor(world, pos);
+        if (CurrentModConfig.burnable && CurrentModConfig.burnableCatchFire && color.equals(BURNABLE_COLOR)) {
+            return BURNABLE_FLAMMABILITY;
+        }
+        return 0;
+    }
+
+    @Override
+    public int getFireSpreadSpeed(IBlockAccess world, BlockPos pos, EnumFacing face) {
+        EnumColor color = getColor(world, pos);
+        if (CurrentModConfig.burnable && CurrentModConfig.burnableCatchFire && color.equals(BURNABLE_COLOR)) {
+            return BURNABLE_FIRE_SPREAD_SPEED;
+        }
+        return 0;
+    }
 }