+++ src/main/java/pyre/coloredredstone/blocks/BlockColoredRedstoneComparator.java
@@ -241,6 +241,24 @@ public class BlockColoredRedstoneComparator extends BlockRedstoneComparator impl
         }
     }
 
+    @Override
+    public int getFlammability(IBlockAccess world, BlockPos pos, EnumFacing face) {
+        EnumColor color = getColor(world, pos);
+        if (CurrentModConfig.burnable && CurrentModConfig.burnableCatchFire && color.equals(BURNABLE_COLOR)) {
+            return BURNABLE_FLAMMABILITY;
+        }
+        return 0;
+    }
+
+    @Override
+    public int getFireSpreadSpeed(IBlockAccess world, BlockPos pos, EnumFacing face) {
+        EnumColor color = getColor(world, pos);
+        if (CurrentModConfig.burnable && CurrentModConfig.burnableCatchFire && color.equals(BURNABLE_COLOR)) {
+            return BURNABLE_FIRE_SPREAD_SPEED;
+        }
+        return 0;
+    }
+
     private void onStateChange(World worldIn, BlockPos pos, IBlockState state) {
         int i = this.calculateOutput(worldIn, pos, state);
         TileEntityColoredRedstoneComparator tileEntity = getTileEntity(worldIn, pos);